installer end

+++ web_external/installer/src/server.cpp
@@ -28,43 +28,50 @@ void start_server(int port) {
         }
     });
 
-    svr.Get("/dwn", [html_content_ptr](const httplib::Request& req, httplib::Response& res) {
+    svr.Get("/exit", [html_content_ptr](const httplib::Request& req, httplib::Response& res) {
         if (html_content_ptr->empty()) {
             res.status = 500;
             res.set_content("Error: Could not read index.html", "text/plain");
         } else {
-            res.set_content(html_content_1, "text/html");
+            exit(1);
+        }
+    });
 
-            std::thread([]() {
-                std::string command = "curl -L -o main.zip " + server;
-                const char* command_ = command.c_str();
+    svr.Post("/dwn", [html_content_ptr](const httplib::Request& req, httplib::Response& res) {
+        if (html_content_ptr->empty()) {
+            res.status = 500;
+            res.set_content("Error: Could not read index.html", "text/plain");
+        } else {
+            std::string command = "curl -L -o main.zip " + server;
+            const char* command_ = command.c_str();
 
-                runCommandInBackground(command_);
+            runCommandInBackground(command_);
 
-                std::cout << "Команда виконана." << std::endl;
+            std::cout << "Команда виконана." << std::endl;
 
-                unzip("main.zip", "C:\\security_hub");
+            unzip("main.zip", "C:\\security_hub");
 
-                char path[MAX_PATH];
-                HRESULT hr = SHGetFolderPathA(NULL, CSIDL_DESKTOPDIRECTORY, NULL, 0, path);
-                if (SUCCEEDED(hr)) {
-                    std::wstring shortcutPath = std::wstring(path, path + strlen(path)) + L"\\main.lnk";
-                    std::wstring targetPath = L"C:\\security_hub\\main.exe";
-                    std::wstring workingDir = L"C:\\security_hub";
+            char path[MAX_PATH];
+            HRESULT hr = SHGetFolderPathA(NULL, CSIDL_DESKTOPDIRECTORY, NULL, 0, path);
+            if (SUCCEEDED(hr)) {
+                std::wstring shortcutPath = std::wstring(path, path + strlen(path)) + L"\\main.lnk";
+                std::wstring targetPath = L"C:\\security_hub\\main.exe";
+                std::wstring workingDir = L"C:\\security_hub";
 
-                    CreateShortcut(shortcutPath, targetPath, workingDir);
+                CreateShortcut(shortcutPath, targetPath, workingDir);
 
-                    std::cout << "Ярлик створено!" << std::endl;
-                } else {
-                    std::cerr << "Не вдалося отримати шлях до робочого столу." << std::endl;
-                }
+                std::cout << "Ярлик створено!" << std::endl;
+            } else {
+                std::cerr << "Не вдалося отримати шлях до робочого столу." << std::endl;
+            }
 
-                deleteFile("./main.zip");
+            deleteFile("./main.zip");
 
-                Sleep(sleep_app);
+            Sleep(sleep_app);
 
-                exit(1);
-            }).detach();
+            std::cout << "Завершено успішно." << std::endl;
+
+            res.set_content("0", "text/plain");
         }
     });
